Typescript Support Added
adding index.d.ts file is not sufficient to work in typescript. Its required to add ts-node support In order to work in typescript files. Also jasmine test added to utilise TDD. Finally new include added as AEHelper.ts ( which is complied to with ae.create function and add as jsx file. That helper responsible for converting node js file path and plaining to use for item finder notationn